smaller changes, implemented selection of single raw data record for evaluations
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@160 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ qfrawdatarecord.h
@@ -280,7 +280,7 @@ class QFRawDataRecord : public QObject {
         };
         /** \brief get the i-th evaluation name */
         inline QString resultsGetEvaluationName(unsigned int i) {
-            if (i<results.size()) return results.keys().at(i);
+            if ((long)i<results.size()) return results.keys().at(i);
             return QString("");
         };
         /** \brief get the i-th result name */
